+++ src/tweaksConfig.ts
@@ -1,4 +1,22 @@
 /**
+ * The different reasons when the video codec should be reused.
+ */
+export enum ForceReuseVideoCodecReason {
+  /**
+   * The new video quality color information is not compatible.
+   */
+  ColorInfoMismatch = 'ColorInfoMismatch',
+  /**
+   * The new video quality exceed the decoder's configured maximum sample size.
+   */
+  MaxInputSizeExceeded = 'MaxInputSizeExceeded',
+  /**
+   * The new video quality exceed the decoder's configured maximum resolution.
+   */
+  MaxResolutionExceeded = 'MaxResolutionExceeded',
+}
+
+/**
  * This configuration is used as an incubator for experimental features. Tweaks are not officially
  * supported and are not guaranteed to be stable, i.e. their naming, functionality and API can
  * change at any time within the tweaks or when being promoted to an official feature and moved
@@ -170,4 +188,14 @@ export interface TweaksConfig {
    * @platform iOS
    */
   updatesNowPlayingInfoCenter?: boolean;
+
+  /**
+   * When switching between video formats (eg: adapting between video qualities) the codec might be recreated due to several reasons.
+   * This behaviour can cause brief black screens when switching between video qualities as codec recreation can be slow.
+   *
+   * Default is `[]` i.e not set
+   *
+   * @platform Android
+   */
+  forceReuseVideoCodecReasons?: Array<ForceReuseVideoCodecReason>;
 }